Added Array.prototype.forEvery
Array method that takes either 2 or 3 arguments. Takes a number as its first argument, and optionally, true, as its second argument. The third argument is a callback function that will be run on every n-th item on the array, where n is the first argument. If true is omitted, the callback function should be the second argument. If true is included, the method will return all items from the original array, even if they are not mutated. If omitted, the method will only return the items on which the callback function was used.